Release lock if the originating process is gone
This fix checks if the IP of the current machine is the same that the IP of the server that acquired the lock. if we are on the same machine, but the originating process doesn't exists anymore, it assumes we can safely release the lock and return false (aka, lock is not locked).

Caveat : If the lock has been acquired by a process on pc-01, but the process failed to release the lock, no other machine will be able to acquire the same lock until a new process on pc-01 will call the isLocked() method again.